[refactor] - 복잡한 코드 분리
